AppRun: Sort library symbols
Currently the AppRun binary does not sort the symbols it finds. This
causes it to sometimes consider a library to be an older version than it
actually is.

Replace the binary with a patched one from
lat9nq/AppImageKit-checkrt@982633f9aecdd490e5bd8b5dcd3203be38414b48
which does sort the symbols before trying to replace it.